requirements.txt
asgiref==3.4.1
astroid==2.11.6
asttokens==2.0.5
attrs==22.1.0
backcall==0.2.0
black==22.6.0
click==8.0.3
colorama==0.4.5
cssbeautifier==1.14.6
decorator==5.1.1
dill==0.3.5.1
Django==3.2.9
django-crispy-forms==1.14.0
django-debug-toolbar==3.2.4
django-environ==0.8.1
django-extensions==3.1.5
django-filter==21.1
djangorestframework==3.12.4
djlint==1.17.2
EditorConfig==0.12.3
executing==0.8.3
html-tag-names==0.1.2
html-void-elements==0.1.0
importlib-metadata==4.12.0
iniconfig==1.1.1
ipython==8.2.0
isort==5.10.1
jedi==0.18.1
jsbeautifier==1.14.6
lazy-object-proxy==1.7.1
Markdown==3.3.4
matplotlib-inline==0.1.3
mccabe==0.7.0
mypy-extensions==0.4.3
mysqlclient==2.1.0
numpy==1.21.3
packaging==21.3
pandas==1.3.4
parso==0.8.3
pathspec==0.10.1
pep517==0.12.0
pexpect==4.8.0
pickleshare==0.7.5
Pillow==9.0.1
pip-tools==6.4.0
platformdirs==2.5.2
pluggy==1.0.0
prompt-toolkit==3.0.29
ptyprocess==0.7.0
pure-eval==0.2.2
py==1.11.0
Pygments==2.11.2
pylint==2.14.3
pyparsing==3.0.9
pytest==7.1.3
pytest-django==4.5.2
python-dateutil==2.8.2
pytz==2021.3
PyYAML==6.0
regex==2022.9.13
reportlab==3.6.6
six==1.16.0
sqlparse==0.4.2
stack-data==0.2.0
tomli==2.0.1
tomlkit==0.11.0
tqdm==4.64.1
traitlets==5.1.1
typing_extensions==4.2.0
wcwidth==0.2.5
whitenoise==5.3.0
wrapt==1.14.1
zipp==3.8.1
